+++ typings/custom-typings.d.ts
@@ -2,11 +2,14 @@ declare module 'rc-switch';
 declare module 'key';
 declare module 'jszip';
 declare module 'clipboard';
+declare module '*.css';
+declare module '*.less';
 
-declare var TEMPLATE;
-declare var RepoMeta;
-declare var Components;
+declare let TEMPLATE;
+declare let RepoMeta;
+declare let Components;
 
 interface Window {
   _gaq: any[];
+  RepoMeta: any;
 }