+++ CHANGELOG.md
@@ -4,11 +4,14 @@
 ### Bug Fixes
 
 * avoid Chrome StorageArea illegal invocation after Manifest V3 upgrade ([e99ee86](https://github.com/ineo6/git-master/commit/e99ee86))
+* restore extension options and popup rendering after React 19 upgrade
+* keep legacy jQuery plugins compatible with jQuery 4
 
 
 ### Features
 
 * upgrade Chrome extension output to Manifest V3 ([e99ee86](https://github.com/ineo6/git-master/commit/e99ee86))
+* upgrade frontend and build tooling dependencies
 
 ## [1.17.1](https://github.com/ineo6/git-master/compare/v1.17.0...v1.17.1) (2022-01-24)